+++ .changes/api-Resource.md
@@ -1,0 +1,5 @@
+---
+'@tauri-apps/api': 'patch:feat'
+---
+
+Exposed `Resource` class which should be extended for Rust-backed resources created through `tauri::Manager::resources_table`.
