+++ .changes/tauri-resources-table.md
@@ -2,4 +2,4 @@
 'tauri': 'patch:feat'
 ---
 
-Exposed `Manager::resources_table` to access the resources table used by tauri, which could be used by plugins or app authors to store their resources and retrieve it later using an id.
+Exposed `Manager::resources_table` to access the resources table used by tauri, which could be used by plugins or app authors to store their resources and retrieve it later using an id and can be used to create Rust-backed resources in JS.
