--- tooling/api/src/internal/index.ts
@@ -1,34 +1,0 @@
-// Copyright 2019-2023 Tauri Programme within The Commons Conservancy
-// SPDX-License-Identifier: Apache-2.0
-// SPDX-License-Identifier: MIT
-
-import { invoke } from '../core'
-
-/**
- * A rust-backed resource.
- *
- * The resource lives in the main process and does not exist
- * in the Javascript world, and thus will not be cleaned up automatiacally
- * except on application exit. If you want to clean it up early, call {@linkcode Resource.close}
- */
-export class Resource {
-  readonly #rid: number
-
-  get rid(): number {
-    return this.#rid
-  }
-
-  constructor(rid: number) {
-    this.#rid = rid
-  }
-
-  /**
-   * Destroys and cleans up this resource from memory.
-   * **You should not call any method on this object anymore and should drop any reference to it.**
-   */
-  async close(): Promise<void> {
-    return invoke('plugin:resources|close', {
-      rid: this.rid
-    })
-  }
-}
